Добавление библиотеки
